fix: erro pagina 404

+++ services/payment-service/src/main/java/com/ordersystem/payment/service/PaymentService.java
@@ -35,7 +35,7 @@ public class PaymentService {
             orderEvent.getOrderId(),
             paymentId,
             paymentStatus,
-            orderEvent.getTotalAmount(),
+            orderEvent.getTotalAmount().doubleValue(),
             LocalDateTime.now()
         );
